Adds compiling aws package, broken tests, and haven't seperated modules
