+++ packages/aws/dynamo_db.go
@@ -1,0 +1,9 @@
+package aws
+
+import (
+	adynamo "github.com/aws/aws-sdk-go/service/dynamodb"
+)
+
+type DynamoDbInterface interface {
+	GetItem(*adynamo.GetItemInput) (*adynamo.GetItemOutput, error)
+}
