+++ packages/aws/go.mod
@@ -1,0 +1,8 @@
+module theodo.red/creditcompanion/packages/aws
+
+go 1.15
+
+require (
+	github.com/aws/aws-sdk-go v1.36.31
+	github.com/stretchr/testify v1.7.0
+)
