+++ packages/tokens/repositories/mocks.go
@@ -1,27 +1,9 @@
 package repositories
 
 import (
-	"github.com/aws/aws-sdk-go/service/dynamodb"
 	"github.com/stretchr/testify/mock"
 	"theodo.red/creditcompanion/packages/tokens/models"
 )
-
-type DynamoDbMock struct {
-	mock.Mock
-}
-
-func (m *DynamoDbMock) GetItem(input *dynamodb.GetItemInput) (*dynamodb.GetItemOutput, error) {
-	args := m.Called(input)
-	returnValue, ok := args.Get(0).(*dynamodb.GetItemOutput)
-	if !ok {
-		return nil, args.Error(1)
-	}
-	return returnValue, args.Error(1)
-}
-
-func testDynamoDbMockImplementsInterface() DynamoDbInterface {
-	return new(DynamoDbMock)
-}
 
 type TokenRepositoryMock struct {
 	mock.Mock
